+++ src/tests.rs
@@ -1,7 +1,7 @@
 use core::{cell::Cell, f32};
 use std::thread_local;
 
-use crate::{Constant, Noise, NoiseFn, OpenSimplex2, Sample};
+use crate::{Constant, Noise, NoiseFn, OpenSimplex2, Perlin, Sample};
 
 #[test]
 fn translate() {
@@ -140,4 +140,13 @@ where
     if max_error > epsilon {
         panic!("error too big!");
     }
+}
+
+#[test]
+fn test_trait_object_noise() {
+    let perlin_dyn: &dyn Sample<2> = &Perlin;
+    let _value = (&perlin_dyn).sample2([1.0, 2.0]);
+
+    let perlin_box_dyn: alloc::boxed::Box<dyn Sample<2>> = alloc::boxed::Box::new(Perlin);
+    let _value = perlin_box_dyn.sample2([1.0, 2.0]);
 }